feat: Add autocomplete functionality to ship command
This commit adds autocomplete functionality to the ship command in order to provide a better user experience. The command now responds with a list of manufacturers or models based on the user's input. The autocomplete feature is triggered when the user starts typing a manufacturer or model name. The list of options is filtered and limited to the top 25 matches. The ship information is retrieved from a Google Spreadsheet using the Google service.